+++ App/templates/addAssessment.html
@@ -9,7 +9,7 @@
 {% block content %}
 
 <div class="formContainer">
-    <form id="addAssessmentForm" action="">
+    <form id="addAssessmentForm" action="/addNewAsg" method="POST">
         <div class="formTitle">
             <h2>Add Assessment</h2>
         </div>
