+++ App/templates/modifyAssessment.html
@@ -8,7 +8,7 @@
 
 {% block content %}
 <div class="formContainer">
-    <form id="modifyAssessmentForm" action="">
+    <form id="modifyAssessmentForm" action="/updateAsg" method="POST">
         <div class="formTitle">
             <h2>Modify Assessment</h2>
         </div>
